requirements.txt
# 암호화폐 거래소 API
ccxt>=4.1.13

# 데이터 분석 및 수치 계산
pandas>=2.1.4
numpy>=1.26.2

# 시각화
matplotlib>=3.8.2
seaborn>=0.13.0

# 설정 파일 처리
pyyaml>=6.0.1

# 텔레그램 알림
python-telegram-bot>=20.7

# 테스트
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# 노트북
jupyter>=1.0.0
notebook>=7.0.6
ipykernel>=6.27.0

# 데이터 처리 및 분석
scikit-learn>=1.3.2
ta>=0.11.0  # 기술적 분석 지표
statsmodels>=0.14.0  # 통계 분석

# 유틸리티
python-dotenv>=1.0.0  # 환경 변수 관리
requests>=2.31.0  # HTTP 요청
websockets>=12.0  # 실시간 데이터 수신
aiohttp>=3.9.1  # 비동기 HTTP 클라이언트
schedule>=1.2.1  # 작업 스케줄링

# 로깅 및 모니터링
structlog>=24.1.0  # 구조화된 로깅
prometheus-client>=0.19.0  # 메트릭 수집

# 개발 도구
black>=23.11.0  # 코드 포맷팅
flake8>=6.1.0  # 코드 린팅
mypy>=1.7.1  # 타입 체크
isort>=5.12.0  # import 정렬

# 기본 패키지
numpy>=1.26.2
pandas>=2.2.1
python-dotenv>=1.0.1

# 데이터 분석 및 머신러닝
scikit-learn>=1.3.2
statsmodels>=0.14.0
scipy>=1.10.0
textblob>=0.17.1

# 기술적 분석
ta-lib>=0.6.3
ta>=0.11.0

# 차트 및 시각화
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
mplfinance>=0.12.9b7  # 최신 베타 버전 사용

# 웹 대시보드
dash>=2.9.0
dash-bootstrap-components>=1.4.0
flask>=2.2.0
streamlit>=1.32.0

# 데이터베이스
peewee>=3.16.0
sqlalchemy>=2.0.0

# API 및 네트워킹
aiohttp>=3.9.3
requests>=2.31.0
websockets>=12.0
ccxt>=4.2.15
aiodns>=3.1.1

# 암호화 및 보안
cryptography>=41.0.0

# 알림 및 메시징
python-telegram-bot[job-queue]>=20.7

# 유틸리티
PyYAML>=6.0.1
schedule>=1.2.1
pytz>=2023.3
tqdm>=4.65.0

# 로깅 및 모니터링
structlog>=24.1.0
prometheus-client>=0.19.0

# 테스트
pytest>=7.4.3
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# 개발 도구
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0

# 노트북
jupyter>=1.0.0
notebook>=7.0.6
ipykernel>=6.27.0

# 추가된 패키지
streamlit>=1.22.0
plotly>=5.13.0
pandas>=1.5.0
numpy>=1.23.0
ccxt>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
websockets>=10.0
python-telegram-bot>=13.0
SQLAlchemy>=1.4.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.9.0
boto3>=1.26.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0 "streamlit>=1.22.0" 
"plotly>=5.10.0" 
